refactor: move configuration to page components
